Set C binding for unix_readlink_job in its own C file
