+++ src/unix/jbuild
@@ -192,6 +192,7 @@
     unix_rewinddir_job
     unix_readdir_job
     unix_readdir_n_job
+    unix_readlink_job
     not_available
   ))
   (install_c_headers (
