+++ src/unix/lwt_unix_windows.h
@@ -120,7 +120,6 @@ CAMLprim value lwt_unix_system_job(value cmdline)
    | Unavailable primitives                                          |
    +-----------------------------------------------------------------+ */
 
-LWT_NOT_AVAILABLE1(unix_readlink_job)
 LWT_NOT_AVAILABLE3(unix_lockf_job)
 LWT_NOT_AVAILABLE1(unix_gethostname_job)
 LWT_NOT_AVAILABLE1(unix_gethostbyname_job)
