+++ src/unix/windows_c/not_available.c
@@ -74,4 +74,5 @@ LWT_NOT_AVAILABLE1(unix_invalidate_dir)
 LWT_NOT_AVAILABLE1(unix_rewinddir_job)
 LWT_NOT_AVAILABLE1(unix_readdir_job)
 LWT_NOT_AVAILABLE2(unix_readdir_n_job)
+LWT_NOT_AVAILABLE1(unix_readlink_job)
 #endif
